Play nice with caniuse features released in browser versions that don’t exist in the data
When the version is not found in the list of released versions, color nothing

This fixes #2.

Problem for `calc-size` specifically is that the feature is marked as added in Edge 129 but that release is not included in the caniuse`.agents.edge.version_list` because that version – at the time of writing – has not been released yet.

Trying to fall back to the versions declared in BCD instead – which does included Edge 129 – would be possible, but the problem is that it’s an object and the order of the releases is therefore not guaranteed when looping over them.

While I could try and fall back to extracting the release from BCD whose status is marked as "beta" (and the one after that "nightly") and manually push them onto `.agents.edge.version_list` I won’t because (1) caniuse.com itself doesn’t do that and (2) it’s a caniuse data problem.

Instead of trying to fix the data, the workaround is to skip the computation when the `version_added` value doesn’t appear in the `version_list`.